Version 1.5 - bring back to sync with VB code.
Return server-side context and the business object (in an undefined state) from the server to the client in the case of a server-side exception.

Also altered the exception mechanism so it returns the full stack trace data from the original exception back to the client, making debugging of DataPortal_xyz methods much easier.

These changes required addtion of DataPortalException, Server.DataPortalException and CallMethodException classes, and two new entries in CSLA.Resources.Strings.resx.