Enhance the server-side DataPortal to support System.Transactions.TransactionScope style transactions.
To do this the TransactionalAttribute class was extended to accept a type parameter indicating the type of transactional technology to use based on an enum. Also a new TransactionalDataPortal class exists which wraps calls to SimpleDataPortal within a TransactionScope.

The server-side DataPortal class was enhanced to do the proper routing. If the Transactional attribute isn't applied the default behavior is Manual. If it is applied, but without an explicit transaction type the default behavior is to use EnterpriseServices.