Make RolesForProperty a Friend class to restrict its scope.
Add a new GetRolesForProperty overload to return the list of roles as a string array rather than returning the actual List(Of String) - this helps prevent unauthorized changing of the role lists.

Added AccessType enum for the new GetRolesForProperty overload.

Added HasXYZ() methods to make it easier to determine whether roles exist for different access types for each property.

Changed BusinessBase to use the new features in Csla.Security.