Address some issues with ERLB when used for in-place editing in a grid.
Add an EditLevel property to IEditableBusinessObject, and implement that property in BusinessBase. Alter ERLB to change the way it saves items so it drops the edit level to 0, does the save and restores the edit level. Also addressed some issues around possible multiple saves if the edit level was >1, and removed an extraneous Insert call after each Delete.