Fix some issues with SetItem and n-level undo behaviors for children in lists.
SetItem now properly configures the newly added item by ensuring that its EditLevel and EditLevelAdded values are both correct.

When a child is deleted from the list, its edit level is lowered to match the list (if necessary). This resolves an issue with data binding in a grid, where the child is the currently selected item in the grid (because its edit level would have been elevated in that case).

In BusinessBase, EditLevelAdded is now NotUndoable, which it should have been all along. This change means that the value is controlled as expected.