Contains fixes for issue #'s 388, 390 and 391.
Issue 388: Added a property to Csla.Serialization.Mobile.SerializationInfo.FieldData, used to indicate the type of the enum stored in the object if any. This is required for properly deserializing enums without adding their types to the data contract's known types (which is not possible). Enums are converted to their integer base type prior to serialization and converted back using the type name upon deserialization.

Issue 390: New class called Csla.Serialization.Mobile.ReferenceComparer<T> was added. Is an implementation of IEqualityComparer<T> that uses Object.ReferenceEquals() to always do a reference-based comparison regardless of whether objects do or do not override Equals(). ReferenceComparer is then provided to the MobileFormatter._serializationReferences dictionary constructor as the default comparer.

Issue 391: New class called Csla.Serialization.Mobile.NullPlaceholder was added. Implements IMobileObject and is used as a placeholder for null child objects in MobileFormatter._serializationReferences. Deserialization logic was added checking for NullPlaceholder, and substituting null for the type instead.

bugid: 388,390,391