Major rework of PropertyStatus in this regard:
There's now a Property property, which should be bound to the business object property. It also supports dot notation to drill through a ViewModel or data provider, so the Path part can be like "Data.Name" or "Model.Name".

There's now a TargetControl property, which should be element bound to the target UI control that is to be managed by PropertyStatus.
bugid: 492