#345 Merge Silverlight code files into the main Csla code
A number of code files (in whole or part) existed in the
Csla.Silverlight folder structure and were linked to other projects.

This is the start of merging those code files into the main Csla folder
structure and then relinking the other projects to the main Csla project
folder.

At this point I've merged the files and have the main Csla project
compiling at least.