#345 Update code files/links in Csla.WinRT.Phone and fix MVC references
Updated the code files and file links in the Csla.WinRT.Phone project so
all the core Csla projects now build. The phone project now references
nearly all the files in the Csla core project, so these changes also
move closer to the goal of using a shared project.

Updated the NuGet package references in the MVC 4 and MVC 5 projects so
they build correctly now in VS15.