#539 Update IsSerialized method to work properly
Notice the string compare for the typename. This is because I'm
sometimes getting the Serializable attribute from CSLA, and sometimes
from mscorlib (even though there is no such attribute in the mscorlib
public surface area). To overcome this issue I've switched to a string
comparison rather than relying on the actual Type.