#712 Add Identity property to IBusinessObject
Add Identity property to IBusinessObject so all objects will now carry a unique identifier (within the scope of the object graph).

Also add code to IParent to generate unique identity values within an object graph.

Still need to wire up calls to GetNextIdentity as objects are created or added to the graph.

Still need to traverse the graph on deserialization to set the _nextIdentity value - or add that value to the serialization stream, but I think it will be best to recalculate it.