`Update SaveState method in StateManager.cs`
Refactored the `SaveState` method in `StateManager.cs`. Removed the calls to `SendSession` method of `_sessionManager` that were previously using `timeout` and `ct` parameters. The `SaveState` method no longer depends on `SendSession` method of `_sessionManager`.

Overload InitializeAsync and refactor StateManager.cs

Overloaded the `InitializeAsync()` method in `StateManager.cs` to accept a `CancellationToken` parameter and added a new private method `GetState(CancellationToken ct)`. This new method checks if the operating system is a browser and retrieves the session using the provided cancellation token. Removed the previous versions of `GetState(CancellationToken ct)` and `InitializeAsync(CancellationToken ct)` methods that were located in a different part of the code. Also, modified the `SaveState(CancellationToken ct)` method by removing the code block that sends the session using the provided cancellation token when the operating system is a browser.