Added async methods and updated parameters in NoOpSessionManager.cs
In NoOpSessionManager.cs, two new asynchronous methods `RetrieveSession(CancellationToken ct)` and `SendSession(CancellationToken ct)` have been added. These methods are versions of the existing `RetrieveSession` and `SendSession` methods that take a `CancellationToken` as a parameter. The `UpdateSession` method has also been modified to include a new parameter `newSession`.

In ViewModel.cs, unnecessary using directives have been removed and a string literal in `SaveAsync` method has been replaced with `nameof(SaveAsync)` for better refactoring resilience.

In BusyHelper.cs, a line of code in the `WaitForIdle` method has been modified, but the change does not seem to alter the functionality. This could be a mistake or an unnecessary change.