Refactor authorization rules and update related files
In this commit, we've made significant changes to the authorization rules and related files. The `using Csla.Core;` statement was removed from `AuthTests.cs` and the `NoAuthAsync` class was refactored to inherit from `AuthorizationRuleAsync` instead of `IAuthorizationRuleAsync`. The `Element`, `Action`, and `CacheResult` properties were removed from `NoAuthAsync` and a constructor was added that takes an `AuthorizationActions` parameter. The `ExecuteAsync` method was also changed to be `protected` and `override`.

The `IAuthorizationRule` and `IAuthorizationRuleAsync` interfaces were removed from `IAuthorizationRuleBase.cs`. A new file `AuthorizationRuleAsync.cs` was added, which defines an abstract class `AuthorizationRuleAsync` that implements `IAuthorizationRuleAsync`. This class includes properties for `CacheResult`, `Element`, and `Action`, as well as a `CanWriteProperty` method and a `ReadProperty` method.

New files `IAuthorizationRule.cs` and `IAuthorizationRuleAsync.cs` were added, which define interfaces for authorization rules. These interfaces include a `Execute` or `ExecuteAsync` method, respectively.