Update project to .NET 8.0 and refactor code
Updated the target framework for the projects from .NET 6.0 and .NET 5.0 to .NET 8.0. This includes changes to the `.csproj` files and the conditional compilation symbols in the code. Package references have been updated to their respective versions for .NET 8.0. Removed code blocks and package references specific to .NET 6.0 or .NET Core 3.1. Introduced the use of `AssemblyLoadContext` class for managing the execution context of an assembly. Updated the usage of `System.Runtime.Loader` and `Csla.Runtime` namespaces, type rules and method cache handling, transaction types handling, and `TransactionalTypes` enum to be compatible with .NET 8.0. Updated usage of several classes and `AssemblyLoadContextManager` based on the .NET version.