Add unit test for managed backing fields analysis
A new test method `AnalyzeWhenClassHasManagedBackingFieldNotStatic` was added to the `EvaluateManagedBackingFieldsAnalayzerTests` class. This method is marked with the `[TestMethod]` attribute and is asynchronous, returning a `Task`. It defines a C# class `A` with various properties and `PropertyInfo` fields to demonstrate different uses of managed backing fields. The method calls `TestHelpers.RunAnalysisAsync` with the `EvaluateManagedBackingFieldsAnalayzer` type and a list of three analyzer identifiers to run the analysis on the provided code. The new test method is positioned before the existing `AnalyzeWhenClassHasManagedBackingFieldUsedProperty` test method.