Simplify LocalProxyOptions initialization in UseLocalProxy (#4195)
The method `UseLocalProxy` now directly initializes a new
`LocalProxyOptions` instance without checking for an existing
instance. The conditional logic that checked for an existing
`IDataPortalProxy` service and its implementation instance has
been removed. This change simplifies the method by ensuring
that a new `LocalProxyOptions` instance is always created and
used.