#4200
Remove ICheckRoles, ICslaIdentity, and ICslaPrincipal interfaces

The `ICheckRoles.cs` file has been completely removed. This file
contained an interface `ICheckRoles` which defined a method
`IsInRole` to check if the current user is in a specified security
role.

The `ICslaIdentity.cs` file has been modified to remove the
interface `ICslaIdentity`. This interface extended `IReadOnlyBase`,
`IIdentity`, and `ICheckRoles`, and included a property `Roles`
which was a list of roles for the user.

The `ICslaPrincipal.cs` file has been modified to remove the
interface `ICslaPrincipal`. This interface extended `IPrincipal`
and `Serialization.Mobile.IMobileObject`.