Features/remove obsolete classes (#4202)
* Remove DataPortalProxyDescriptor and IApplicationContextProvider

The DataPortalProxyDescriptor class in the Csla.DataPortalClient
namespace has been completely removed. This class included
properties for ProxyTypeName and DataPortalUrl.

Additionally, the IApplicationContextProvider interface in the
Csla.Channels.Local namespace, which was marked as internal,
has also been removed.

* #4199

Remove PrincipalCache and related configurations

Removed the `PrincipalCache` class and its associated file `PrincipalCache.cs`, which provided caching for principal objects at the AppDomain level. Also removed the `PrincipalCacheMaxCacheSize` property from the `SecurityOptions` class in `SecurityOptions.cs`, eliminating the configuration for the maximum cache size.

* #4200

Remove ICheckRoles, ICslaIdentity, and ICslaPrincipal interfaces

The `ICheckRoles.cs` file has been completely removed. This file
contained an interface `ICheckRoles` which defined a method
`IsInRole` to check if the current user is in a specified security
role.

The `ICslaIdentity.cs` file has been modified to remove the
interface `ICslaIdentity`. This interface extended `IReadOnlyBase`,
`IIdentity`, and `ICheckRoles`, and included a property `Roles`
which was a list of roles for the user.

The `ICslaPrincipal.cs` file has been modified to remove the
interface `ICslaPrincipal`. This interface extended `IPrincipal`
and `Serialization.Mobile.IMobileObject`.