#4622 Implement virtual Deserialized method (#4744)
* #4622 Implement virtual Deserialized method

FIRST:
Can the method ISerializationNotification.Deserialized be marked overridable so we can hook into it?...as a replacement for the old OnDeserialized. No need for any parameters in the new method AFAIK.

SECOND:
See ReadOnlyBase.cs Line 513 -> Inside the method void ISerializationNotification.Deserialized(), why does the code new up an System.Runtime.Serialization.StreamingContext() and pass it to the method private void OnDeserializedHandler(? That context parameter is unused so it is not needed. Also that extra method seems pointless, the little code there could simply be moved into original method.
Same thing in BusinessRules.cs

THIRD:
Why is the attribute [System.Runtime.Serialization.OnDeserialized] added to the previously mentioned (seemingly useless) method private void OnDeserializedHandler only in the ReadOnlyBase and BusinessRules classes? That attribute is not used or referenced anywhere.

* Restore accidentally removed code